refactor(quotes): move Opendax driver to a separate package
