+++ pkg/quotes/opendax/protocol/args_iterator.go
@@ -1,4 +1,4 @@
-package opendax_protocol
+package protocol
 
 import (
 	"errors"
