+++ pkg/quotes/opendax/protocol/parser_test.go
@@ -1,4 +1,4 @@
-package opendax_protocol
+package protocol
 
 import (
 	"testing"
