+++ pkg/quotes/opendax/protocol/protocol.go
@@ -1,5 +1,5 @@
 // Events/metods names for open-finance.org protocol implementation client and server.
-package opendax_protocol
+package protocol
 
 import "fmt"
